backend redis done (needs unit tests)
