+++ src/geo/etl/aggieExpertsAPI/redis/redisUtils.js
@@ -13,7 +13,16 @@
 const { cacheExperts } = require('./expertCache');
 const { cacheGrants } = require('./grantCache');
 const { cacheWorks } = require('./workCache');
-const { createRedisClient, sanitizeString } = require('./redisConfig');
+const { createRedisClient } = require('./redisConfig');
+
+// Helper function to sanitize strings
+function sanitizeString(input) {
+  if (!input) return '';
+  return input
+    .replace(/[^\w\s.-]/g, '') // Remove special characters except word characters, spaces, hyphens, and periods
+    .replace(/\s+/g, ' ')      // Replace multiple spaces with a single space
+    .trim();                   
+}
 
 /**
  * Check if Redis is available
